Show a read more button when the post is truncated on the home page.
This required to fix the theme, as the read more button was shown also for
the first post even though it was not truncated. That happened because for
some reason the length of the content was always computed to be 1 character
longer than the summary, so the .Truncated variable was true.